+++ internal/command/darwin.go
@@ -17,9 +17,9 @@ const (
 
 var (
 	// darwinArchSupported defines the supported target architectures on darwin
-	darwinArchSupported = []Architecture{ArchAmd64, Arch386}
+	darwinArchSupported = []Architecture{ArchAmd64, ArchArm64}
 	// darwinImage is the fyne-cross image for the Darwin OS
-	darwinImage = "fyneio/fyne-cross:base-latest"
+	darwinImage = "fyneio/fyne-cross:darwin-latest"
 )
 
 // Darwin build and package the fyne app for the darwin OS
@@ -209,9 +209,9 @@ func darwinContext(flags *darwinFlags, args []string) ([]Context, error) {
 
 		switch arch {
 		case ArchAmd64:
-			ctx.Env = append(ctx.Env, "GOOS=darwin", "GOARCH=amd64", "CC=o32-clang")
-		case Arch386:
-			ctx.Env = append(ctx.Env, "GOOS=darwin", "GOARCH=386", "CC=o32-clang")
+			ctx.Env = append(ctx.Env, "CGO_CFLAGS=-mmacosx-version-min=10.12", "CGO_LDFLAGS=-mmacosx-version-min=10.12", "GOOS=darwin", "GOARCH=amd64", "CC=o64-clang")
+		case ArchArm64:
+			ctx.Env = append(ctx.Env, "GOOS=darwin", "CGO_CFLAGS=-mmacosx-version-min=10.12", "CGO_LDFLAGS=-mmacosx-version-min=10.12", "GOARCH=arm64", "CC=o64-clang")
 		}
 
 		// set context based on command-line flags