+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Fix android keystore path is not resolved correctly
+- Fix some release flags are always set even if empty
 
 ## [1.0.0]
 - Add support for "fyne release" #3