+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog - Fyne.io fyne-cross
 
 ## Unreleased
+- Update Go to v1.16.0
+- Update fyne cli to v2.0.1
+- Add darwin arm64 target #33
+- Remove darwin 386 target
 - Fix android keystore path is not resolved correctly
 - Fix some release flags are always set even if empty
 - Fix appID flag should not have a default #25